requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
auto-py-to-exe==2.46.0
beautifulsoup4==4.13.3
bidict==0.23.1
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dash==3.0.2
distro==1.9.0
Eel==0.18.1
et_xmlfile==2.0.0
Flask==3.0.3
Flask-SocketIO==5.5.1
fonttools==4.57.0
future==1.0.0
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.34.0
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.70.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
plotly==6.0.1
prettytable==3.16.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pyecharts==2.0.8
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.3
retrying==1.3.4
seaborn==0.13.2
selenium==4.31.0
setuptools==78.1.0
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tabulate==0.9.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
zhconv==1.4.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
